fix(simulator): ensure allocator consistency during state rollback
