+++ simulator/src/test.rs
@@ -492,7 +492,7 @@ mod contract_execution_tests {
 
         // Create empty operations vector
         let operations: VecM<Operation, 100> = VecM::default();
-        let host = soroban_env_host::Host::default();
+        let mut sim_host = crate::runner::SimHost::new(None, None, None);
         let mut coverage = CoverageTracker::default();
 
         // Should succeed with empty operations
@@ -523,7 +523,7 @@ mod contract_execution_tests {
             include_linear_memory: false,
             enable_asset_safety: false,
         };
-        let result = execute_operations(&host, &operations, &request, None, &mut coverage);
+        let result = execute_operations(&mut sim_host, &operations, &request, &mut coverage);
         assert!(result.is_ok());
 
         let logs = result.unwrap();
